requirements.txt
sentence-transformers>=3.0.0
faiss-cpu
google-generativeai>=0.5.0
numpy>=1.24.0
python-dotenv>=1.0.0
